+++ src/controllers/SettingsController.php
@@ -10,6 +10,7 @@ namespace dukt\social\controllers;
 use Craft;
 use craft\web\Controller;
 use dukt\social\Plugin as Social;
+use dukt\social\Plugin;
 use yii\web\Response;
 
 /**
@@ -44,5 +45,37 @@ class SettingsController extends Controller
         $variables['totalAccounts'] = count($accounts);
 
         return $this->renderTemplate('social/settings/settings', $variables);
+    }
+
+    /**
+     * Saves the settings.
+     *
+     * @return null|Response
+     * @throws \yii\web\BadRequestHttpException
+     */
+    public function actionSaveSettings()
+    {
+        $this->requirePostRequest();
+        $settings = Craft::$app->getRequest()->getBodyParam('settings', []);
+        $plugin = Craft::$app->getPlugins()->getPlugin('social');
+
+        if ($plugin === null) {
+            throw new NotFoundHttpException('Plugin not found');
+        }
+
+        if (!Plugin::getInstance()->savePluginSettings($settings, $plugin)) {
+            Craft::$app->getSession()->setError(Craft::t('app', 'Couldn’t save plugin settings.'));
+
+            // Send the plugin back to the template
+            Craft::$app->getUrlManager()->setRouteParams([
+                'plugin' => $plugin
+            ]);
+
+            return null;
+        }
+
+        Craft::$app->getSession()->setNotice(Craft::t('app', 'Plugin settings saved.'));
+
+        return $this->redirectToPostedUrl();
     }
 }